refactor: mocks와 test 폴더 이름 수정
