+++ src/__mocks__/browser.ts
