+++ src/__mocks__/data/products.ts
