+++ src/main.tsx
@@ -10,7 +10,7 @@ import { BASE_URL } from "./shared/config/base.ts";
 async function enableMocking() {
   if (!import.meta.env.VITE_USE_MOCK) return;
 
-  const { worker } = await import("./mocks/browser");
+  const { worker } = await import("./__mocks__/browser.ts");
 
   return worker.start({
     serviceWorker: {
