+++ vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "jsdom",
-    setupFiles: ["./src/__test__/setupTests.ts"],
+    setupFiles: ["./src/__tests__/setupTests.ts"],
     restoreMocks: true,
   },
 });
